Update the demo probe to notify if the health check was quarantined o… (#485)

+++ shell/demo/demo.sh
@@ -53,11 +53,13 @@ function download_test {
 
 function execute_test {
     local _temp=$1
+    local _test_name=$2
     echo -e "\n[ ] Starting test"
     sleep 1 && tput cuu 1 && tput el
     $_temp
     local _res=$?
     if ( echo "100 9 17 18 105 127" | grep -w -q $_res );then
+        [[ ("$_test_name" == 'Health check' || "$_test_name" == 'Will a long running VST be stopped properly?') ]] && [ $_res != 100 ] && echo -e "${YELLOW}[!] Health check should not be quarantined or blocked${NC}"
         echo -e "${GREEN}[✓] Executed test: control test passed${NC}"
     elif [ $_res -eq 101 ];then
         echo -e "${RED}[!] Executed test: control test failed${NC}"
@@ -69,7 +71,7 @@ function execute_test {
 
 function execute_cleanup {
     echo -e "\n[ ] Running cleanup"
-    rm -rf $PRELUDE_DIR/*
+    rm -rf ${PRELUDE_DIR:?}/*
     local _res=$?
     tput cuu 1 && tput el
     if [ $_res -eq 0 ];then
@@ -94,7 +96,7 @@ function run_demo {
     echo -e "\nStarting test at: $(date +"%T")"
     check_relevance
     download_test $_test_id $_temp
-    execute_test $_temp
+    execute_test $_temp "${_test_name}"
     local _res=$?
     execute_cleanup $_temp
     post_results $_test_id $_res
